refactor(design-copilot): resolve icon manifest from workspace, not bundled data
Keep the OSS package free of Árvore-proprietary data. find_icon no longer ships
data/icons.manifest.json or the generator script. Instead it resolves the manifest:

1. ARVORE_ICONS_MANIFEST env var (explicit override), else
2. walk up from cwd looking for design/design-system/icons.manifest.json

The manifest + generator now live in the consuming workspace (arvore-hub).
Removes the tsx devDep and gen:icons script; package ships only dist/.
