+++ packages/design-copilot/package.json
@@ -1,19 +1,17 @@
 {
   "name": "@arvoretech/pi-design-copilot",
   "version": "0.1.0",
-  "description": "PI extension that makes the model much better at Árvore UI work: searchable icon index (find_icon), brand asset registry (find_asset), a mandatory design brief with a hard gate before generating UI, and a heuristics review sub-agent gate.",
+  "description": "PI extension that makes the model much better at Árvore UI work: searchable icon index (find_icon), brand asset registry (find_asset), a mandatory design brief with a hard gate before generating UI, and a heuristics review sub-agent gate. Data (icon manifest, asset registry) lives in the consuming workspace, resolved via ARVORE_ICONS_MANIFEST or a design/design-system path.",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "type": "module",
   "files": [
-    "dist",
-    "data"
+    "dist"
   ],
   "scripts": {
     "build": "tsc",
     "dev": "tsc --watch",
-    "lint": "tsc --noEmit",
-    "gen:icons": "tsx scripts/build-icon-manifest.ts"
+    "lint": "tsc --noEmit"
   },
   "peerDependencies": {
     "@earendil-works/pi-coding-agent": ">=0.74.0",
@@ -26,7 +24,6 @@
     "@earendil-works/pi-tui": "latest",
     "@earendil-works/pi-agent-core": "latest",
     "@types/node": "^20.10.0",
-    "tsx": "^4.7.0",
     "typescript": "^5.3.0"
   },
   "pi": {
